+++ src/Automation/ProMoLabelsAuto.h
@@ -56,7 +56,7 @@ protected:
 	afx_msg VARIANT GetIDs();
 	afx_msg void SetIDs(const VARIANT FAR& newValue);
 	afx_msg LPDISPATCH Add();
-	afx_msg long Count();
+	afx_msg short Count();
 	afx_msg BOOL Remove(const VARIANT FAR& Item);
 	afx_msg LPDISPATCH GetItem(const VARIANT FAR& Item);
 	afx_msg void SetItem(const VARIANT FAR& Item, LPDISPATCH newValue);
