+++ src/Automation/ProMoPropertyAuto.h
@@ -48,8 +48,8 @@ protected:
 	//{{AFX_DISPATCH(CProMoPropertyAuto)
 	afx_msg BSTR GetName();
 	afx_msg void SetName(LPCTSTR lpszNewValue);
-	afx_msg long GetType();
-	afx_msg void SetType(long nNewValue);
+	afx_msg short GetType();
+	afx_msg void SetType(short nNewValue);
 	afx_msg VARIANT GetValue();
 	afx_msg void SetValue(const VARIANT FAR& newValue);
 	afx_msg VARIANT GetChildNames();
@@ -60,7 +60,7 @@ protected:
 	afx_msg BOOL IsMultivalue();
 	afx_msg BOOL Add();
 	afx_msg BOOL Remove();
-	afx_msg long Count();
+	afx_msg short Count();
 	afx_msg LPDISPATCH GetItem(const VARIANT FAR& Item);
 	afx_msg void SetItem(const VARIANT FAR& Item, LPDISPATCH newValue);
 	//}}AFX_DISPATCH
